chore: confine azure blob storage functions
